Separate components from index.jsx ---------------------------------- Plus try to make a onClickListener in digitButtons
