+++ src/components/App.jsx
@@ -1,0 +1,27 @@
+import React, { useState } from 'react';
+
+import Page from './Page';                                                                                                                                                                                                                                                             
+
+//Operate counting & return Page
+function App() {
+    const [state, setState] = useState({
+        count: 0,
+    });
+ 
+    const { count } = state;
+
+    function handleClick() {
+        setState({
+            count: count + 1,
+ });
+    }
+
+    return (
+        <Page
+        count={count}
+        onClick={handleClick}
+        />
+    );
+}
+
+export default App;
