+++ src/components/DigitClickListener.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function DigitClickListener ({ count, onClick }) {
+    return (
+        <div>
+            onClick={onClick}
+        </div>
+    );
+}
+
+export default DigitClickListener;
