+++ src/components/Page.jsx
@@ -1,0 +1,21 @@
+import React from 'react';
+
+import Counter from './Counter';
+import Buttons from './Buttons';
+
+//Drawing UI
+function Page ({ count, onClick }) {
+    return (
+        <div>
+            <p>Counter</p>
+            <Counter 
+                count={count}
+                onClick={onClick}
+            />
+            <Buttons count={count}
+                onClick={onClick}/>
+        </div>
+        );
+}
+
+export default Page;
